Trait imgui::internal::DataTypeKind

pub unsafe trait DataTypeKind: Copy {
    const KIND: DataType;
}

Primitive type marker.

If this trait is implemented for a type, it is assumed to have exactly the same representation in memory as the primitive value described by the associated KIND constant.

Associated Constants

const KIND: DataType

Loading content...