Trait imgui::internal::RawWrapper

pub trait RawWrapper {
    type Raw;
    unsafe fn raw(&self) -> &Self::Raw;

    unsafe fn raw_mut(&mut self) -> &mut Self::Raw;
}

Marks a type as a transparent wrapper over a raw type

Associated Types

type Raw

Wrapped raw type

Loading content...
Required methods

unsafe fn raw(&self) -> &Self::Raw

Returns an immutable reference to the wrapped raw value

unsafe fn raw_mut(&mut self) -> &mut Self::Raw

Returns a mutable reference to the wrapped raw value

Loading content...