Crate imgui

Re-exports

pub extern crate imgui_sys as sys;

Macros

im_str

Structs

ChannelsSplit
Represent the drawing interface within a call to channels_split.

ChildFrame

CollapsingHeader

ColorButton
Builder for a color button widget.

ColorEdit
Builder for a color editor widget.

ColorPicker
Builder for a color picker widget.

DragFloat

DragFloat2

DragFloat3

DragFloat4

DragFloatRange2

DragInt

DragInt2

DragInt3

DragInt4

DragIntRange2

DrawData

DrawList

DrawListIterator

FontGlyphRange
A set of 16-bit Unicode codepoints

FrameSize

ImColor
Wrap ImU32 (a type typically used by ImGui to store packed colors) This type is used to represent the color of drawing primitives in ImGui's custom drawing API.

ImDrawVert
A single vertex

ImFont
A handle to an imgui font.

ImFontAtlas
A handle to imgui's font manager.

ImFontConfig
A builder for the configuration for a font.

ImGui

ImGuiColorEditFlags
Color edit flags

ImGuiCond
Condition flags

ImGuiFocusedFlags
Flags for window focus checks

ImGuiHoveredFlags
Flags for hover checks

ImGuiInputTextFlags
Flags for text inputs

ImGuiSelectableFlags
Flags for selectables

ImGuiStyle
Runtime data for styling/colors

ImGuiTreeNodeFlags
Flags for trees and collapsing headers

ImGuiWindowFlags
Window flags

ImStr

ImString

ImTexture

ImVec2
A tuple of 2 floating-point values

ImVec4
A tuple of 4 floating-point values

Image
Represent an image about to be drawn. See [Ui::image].

InputFloat

InputFloat2

InputFloat3

InputFloat4

InputInt

InputInt2

InputInt3

InputInt4

InputText

InputTextMultiline

Menu

MenuItem

PlotHistogram

PlotLines

PopupModal
Created by call to [Ui::popup_modal].

ProgressBar
Progress bar widget.

SliderFloat

SliderFloat2

SliderFloat3

SliderFloat4

SliderInt

SliderInt2

SliderInt3

SliderInt4

TextureHandle

Textures
Generic texture mapping for use by renderers.

TreeNode

Ui

Window

WindowDrawList
Object implementing the custom draw API.

Enums

ColorEditMode
Color editor mode.

ColorFormat
Color component formatting.

ColorPickerMode
Color picker hue/saturation/value editor mode.

ColorPreview
Color editor preview style.

EditableColor
Mutable reference to an editable color value.

ImGuiCol
A color identifier for styling

ImGuiKey
A key identifier (ImGui-side enum)

ImGuiMouseCursor
A mouse cursor identifier

ImId

ImMouseButton
Represents one of the buttons of the mouse

StyleVar

Functions

get_style_color_name

get_version

Type Definitions

ImDrawIdx
Vertex index