Crate imgui

Re-exports

pub extern crate imgui_sys as sys;

pub use self::color::ImColor32;

pub use self::drag_drop::DragDropFlags;

pub use self::drag_drop::DragDropSource;

pub use self::drag_drop::DragDropTarget;

Modules

color

drag_drop

Structs to create a Drag and Drop sequence. Almost all structs are re-exported and can be accessed from the crate root; some additional utilities can be found in here.

internal

Internal raw utilities (don’t use unless you know what you’re doing!)

Macros

im_str

Structs

AngleSlider

Builder for an angle slider widget.

BackendFlags

Backend capabilities

ChannelsSplit

Represent the drawing interface within a call to channels_split.

ChildWindow

Builder for a child window

ChildWindowToken

Tracks a child window that must be ended by calling .end()

CollapsingHeader

Builder for a collapsing header widget

ColorButton

Builder for a color button widget.

ColorEdit

Builder for a color editor widget.

ColorEditFlags

Color edit flags

ColorPicker

Builder for a color picker widget.

ColorStackToken

Tracks one or more changes pushed to the color stack that must be popped by calling .pop()

ComboBox

Builder for a combo box widget

ComboBoxFlags

Flags for combo boxes

ComboBoxToken

Tracks a combo box that must be ended by calling .end()

ConfigFlags

Configuration flags

Context

An imgui-rs context.

Drag

Builder for a drag slider widget.

DragRange

Builder for a drag slider widget.

DrawCmdIterator

DrawCmdParams

DrawData

All draw data to render a Dear ImGui frame.

DrawList

Draw command list

DrawListIterator

Iterator over draw lists

DrawListMut

Object implementing the custom draw API.

DrawVert

A single vertex

Font

Runtime data for a single font within a font atlas

FontAtlas

A font atlas that builds a single texture

FontAtlasFlags

Font atlas configuration flags

FontAtlasTexture

Handle to a font atlas texture

FontConfig

Configuration settings for a font

FontGlyph

A single font glyph

FontGlyphRanges

A set of 16-bit Unicode codepoints

FontId

A font identifier

FontStackToken

Tracks a font pushed to the font stack that must be popped by calling .pop()

GroupToken

Tracks a layout group that must be ended by calling .end()

IdStackToken

Tracks an ID pushed to the ID stack that must be popped by calling .pop()

ImDrawCornerFlags

Flags for indictating which corner of a rectangle should be rounded

ImDrawListFlags

Draw list flags

ImGuiInputTextFlags

Flags for text inputs

ImStr

A UTF-8 encoded, implicitly nul-terminated string slice.

ImString

A UTF-8 encoded, growable, implicitly nul-terminated string.

Image

Builder for an image widget

ImageButton

Builder for an image button widget

InputFloat

InputFloat2

InputFloat3

InputFloat4

InputInt

InputInt2

InputInt3

InputInt4

InputText

InputTextMultiline

Io

Settings and inputs/outputs for imgui-rs

ItemFlagsStackToken

Tracks a change pushed to the item flags stack

ItemHoveredFlags

Item hover check option flags

ItemWidthStackToken

Tracks a change pushed to the item width stack

ListBox

Builder for a list box widget

ListBoxToken

Tracks a list box that must be ended by calling .end()

ListClipper

MainMenuBarToken

Tracks a main menu bar that must be ended by calling .end()

MenuBarToken

Tracks a menu bar that must be ended by calling .end()

MenuItem

Builder for a menu item.

MenuToken

Tracks a menu that must be ended by calling .end()

PlotHistogram

PlotLines

PopupModal

Created by call to Ui::popup_modal.

ProgressBar

Builder for a progress bar widget.

Selectable

Builder for a selectable widget.

SelectableFlags

Flags for selectables

SharedFontAtlas

A font atlas that can be shared between contexts

Slider

Builder for a slider widget.

SliderFlags

Flags for sliders

Style

User interface style/colors

StyleStackToken

Tracks one or more changes pushed to the style stack that must be popped by calling .pop()

SuspendedContext

A suspended imgui-rs context.

TabBar

Builder for a tab bar.

TabBarFlags

TabBarToken

Tracks a window that must be ended by calling .end()

TabItem

TabItemFlags

TabItemToken

TableBgTarget

TableColumnFlags

TableFlags

Item hover check option flags

TableRowFlags

TextWrapPosStackToken

Tracks a change pushed to the text wrap position stack

TextureId

An opaque texture identifier

Textures

Generic texture mapping for use by renderers.

TooltipToken

Tracks a layout tooltip that must be ended by calling .end()

TreeNode

Builder for a tree node widget

TreeNodeFlags

Flags for tree nodes

TreeNodeToken

Tracks a tree node that must be popped by calling .pop().

Ui

A temporary reference for building the user interface for one frame

VerticalSlider

Builder for a vertical slider widget.

Window

Builder for a window

WindowFlags

Configuration flags for windows

WindowFocusedFlags

Window focus check option flags

WindowHoveredFlags

Window hover check option flags

WindowToken

Tracks a window that must be ended by calling .end()

Enums

ColorEditDisplayMode

Color editor display mode.

ColorEditInputMode

Color editor input mode.

ColorFormat

Color component formatting

ColorPickerMode

Color picker hue/saturation/value editor mode

ColorPreview

Color editor preview style

ComboBoxHeight

Combo box height mode.

ComboBoxPreviewMode

Combo box preview mode.

Condition

Condition for applying a setting

Direction

A cardinal direction

DrawCmd

A draw command

EditableColor

Mutable reference to an editable color value.

FocusedWidget

Target widget selection for keyboard focus

FontAtlasRef

An immutably borrowed reference to a (possibly shared) font atlas

FontAtlasRefMut

A mutably borrowed reference to a (possibly shared) font atlas

FontSource

A source for binary font data

Id

Unique ID used by widgets

ItemFlag

A temporary change in item flags

Key

A key identifier

MouseButton

Represents one of the supported mouse buttons

MouseCursor

Mouse cursor type identifier

NavInput

An input identifier for navigation

StyleColor

A color identifier for styling

StyleVar

A temporary change in user interface style

TreeNodeId

Unique ID used by tree nodes

Traits

ClipboardBackend

Trait for clipboard backends

Functions

dear_imgui_version

Returns the underlying Dear ImGui library version

Type Definitions

DrawIdx

A vertex index

ImGuiTreeNodeFlags