Enum imgui::StyleColor

#[repr(u32)]
pub enum StyleColor {

    Text,
    TextDisabled,
    WindowBg,
    ChildBg,
    PopupBg,
    Border,
    BorderShadow,
    FrameBg,
    FrameBgHovered,
    FrameBgActive,
    TitleBg,
    TitleBgActive,
    TitleBgCollapsed,
    MenuBarBg,
    ScrollbarBg,
    ScrollbarGrab,
    ScrollbarGrabHovered,
    ScrollbarGrabActive,
    CheckMark,
    SliderGrab,
    SliderGrabActive,
    Button,
    ButtonHovered,
    ButtonActive,
    Header,
    HeaderHovered,
    HeaderActive,
    Separator,
    SeparatorHovered,
    SeparatorActive,
    ResizeGrip,
    ResizeGripHovered,
    ResizeGripActive,
    Tab,
    TabHovered,
    TabActive,
    TabUnfocused,
    TabUnfocusedActive,
    PlotLines,
    PlotLinesHovered,
    PlotHistogram,
    PlotHistogramHovered,
    TableHeaderBg,
    TableBorderStrong,
    TableBorderLight,
    TableRowBg,
    TableRowBgAlt,
    TextSelectedBg,
    DragDropTarget,
    NavHighlight,
    NavWindowingHighlight,
    NavWindowingDimBg,
    ModalWindowDimBg,

}

A color identifier for styling

Variants

Text

TextDisabled

WindowBg

Background of normal windows

ChildBg

Background of child windows

PopupBg

Background of popups, menus, tooltips windows

Border

BorderShadow

FrameBg

Background of checkbox, radio button, plot, slider, text input

FrameBgHovered

FrameBgActive

TitleBg

TitleBgActive

TitleBgCollapsed

MenuBarBg

ScrollbarBg

ScrollbarGrab

ScrollbarGrabHovered

ScrollbarGrabActive

CheckMark

SliderGrab

SliderGrabActive

Button

ButtonHovered

ButtonActive

Header

HeaderHovered

HeaderActive

Separator

SeparatorHovered

SeparatorActive

ResizeGrip

ResizeGripHovered

ResizeGripActive

Tab

TabHovered

TabActive

TabUnfocused

TabUnfocusedActive

PlotLines

PlotLinesHovered

PlotHistogram

PlotHistogramHovered

TableHeaderBg

TableBorderStrong

TableBorderLight

TableRowBg

TableRowBgAlt

TextSelectedBg

DragDropTarget

NavHighlight

Gamepad/keyboard: current highlighted item

NavWindowingHighlight

Highlight window when using CTRL+TAB

NavWindowingDimBg

Darken/colorize entire screen behind the CTRL+TAB window list, when active

ModalWindowDimBg

Darken/colorize entire screen behind a modal window, when one is active

Implementations

impl StyleColor

pub const VARIANTS: [StyleColor; 53]

All possible StyleColor variants

pub const COUNT: usize

Total count of StyleColor variants