Enum imgui::StyleVar

pub enum StyleVar {

    Alpha(f32),
    WindowPadding([f32; 2]),
    WindowRounding(f32),
    WindowBorderSize(f32),
    WindowMinSize([f32; 2]),
    WindowTitleAlign([f32; 2]),
    ChildRounding(f32),
    ChildBorderSize(f32),
    PopupRounding(f32),
    PopupBorderSize(f32),
    FramePadding([f32; 2]),
    FrameRounding(f32),
    FrameBorderSize(f32),
    ItemSpacing([f32; 2]),
    ItemInnerSpacing([f32; 2]),
    IndentSpacing(f32),
    ScrollbarSize(f32),
    ScrollbarRounding(f32),
    GrabMinSize(f32),
    GrabRounding(f32),
    TabRounding(f32),
    ButtonTextAlign([f32; 2]),
    SelectableTextAlign([f32; 2]),
    CellPadding([f32; 2]),

}

A temporary change in user interface style

Variants

Alpha(f32)

Tuple Fields

0: f32

Global alpha applies to everything

WindowPadding([f32; 2])

Tuple Fields

0: [f32; 2]

Padding within a window

WindowRounding(f32)

Tuple Fields

0: f32

Rounding radius of window corners

WindowBorderSize(f32)

Tuple Fields

0: f32

Thickness of border around windows

WindowMinSize([f32; 2])

Tuple Fields

0: [f32; 2]

Minimum window size

WindowTitleAlign([f32; 2])

Tuple Fields

0: [f32; 2]

Alignment for title bar text

ChildRounding(f32)

Tuple Fields

0: f32

Rounding radius of child window corners

ChildBorderSize(f32)

Tuple Fields

0: f32

Thickness of border around child windows

PopupRounding(f32)

Tuple Fields

0: f32

Rounding radius of popup window corners

PopupBorderSize(f32)

Tuple Fields

0: f32

Thickness of border around popup/tooltip windows

FramePadding([f32; 2])

Tuple Fields

0: [f32; 2]

Padding within a framed rectangle (used by most widgets)

FrameRounding(f32)

Tuple Fields

0: f32

Rounding radius of frame corners (used by most widgets)

FrameBorderSize(f32)

Tuple Fields

0: f32

Thickness of border around frames

ItemSpacing([f32; 2])

Tuple Fields

0: [f32; 2]

Horizontal and vertical spacing between widgets/lines

ItemInnerSpacing([f32; 2])

Tuple Fields

0: [f32; 2]

Horizontal and vertical spacing between elements of a composed widget (e.g. a slider and its label)

IndentSpacing(f32)

Tuple Fields

0: f32

Horizontal indentation when e.g. entering a tree node

ScrollbarSize(f32)

Tuple Fields

0: f32

Width of the vertical scrollbar, height of the horizontal scrollbar

ScrollbarRounding(f32)

Tuple Fields

0: f32

Rounding radius of scrollbar grab corners

GrabMinSize(f32)

Tuple Fields

0: f32

Minimum width/height of a grab box for slider/scrollbar

GrabRounding(f32)

Tuple Fields

0: f32

Rounding radius of grab corners

TabRounding(f32)

Tuple Fields

0: f32

Rounding radius of upper corners of tabs

ButtonTextAlign([f32; 2])

Tuple Fields

0: [f32; 2]

Alignment of button text when button is larger than text

SelectableTextAlign([f32; 2])

Tuple Fields

0: [f32; 2]

Alignment of selectable text when selectable is larger than text

CellPadding([f32; 2])

Tuple Fields

0: [f32; 2]

Padding within a table cell