Struct imgui::ColorPicker

#[must_use]

pub struct ColorPicker<'ui, 'p> { /* fields omitted */ }

Builder for a color picker widget.

Methods

impl<'ui, 'p> ColorPicker<'ui, 'p>

pub fn new(_: &Ui<'ui>, label: &'p ImStr, value: EditableColor<'p>) -> Self

Constructs a new color picker builder.

pub fn flags(self, flags: ImGuiColorEditFlags) -> Self

Replaces all current settings with the given flags.

pub fn alpha(self, value: bool) -> Self

Enables/disables the use of the alpha component.

pub fn small_preview(self, value: bool) -> Self

Enables/disables the colored square preview next to the inputs.

pub fn inputs(self, value: bool) -> Self

Enables/disables the input sliders/text widgets.

pub fn tooltip(self, value: bool) -> Self

Enables/disables the tooltip that appears when hovering the preview.

pub fn label(self, value: bool) -> Self

Enables/disables display of the inline text label (the label is in any case forwarded to the tooltip and picker).

pub fn side_preview(self, value: bool) -> Self

Enables/disables the bigger color preview on the right side of the picker.

pub fn alpha_bar(self, value: bool) -> Self

Enables/disables the vertical alpha bar/gradient in the color picker.

pub fn preview(self, preview: ColorPreview) -> Self

Sets the preview style.

pub fn rgb(self, value: bool) -> Self

Enables/disables the RGB inputs.

pub fn hsv(self, value: bool) -> Self

Enables/disables the HSV inputs.

pub fn hex(self, value: bool) -> Self

Enables/disables the HEX input.

pub fn mode(self, mode: ColorPickerMode) -> Self

Sets the hue/saturation/value editor mode.

pub fn format(self, format: ColorFormat) -> Self

Sets the formatting style of color components.

pub fn reference_color(self, ref_color: &'p [f32; 4]) -> Self

Sets the shown reference color.

pub fn build(self) -> bool

Builds the color picker.